requirements.txt
scikit-learn
matplotlib
ipykernel
requests
python-dotenv
